Add files via upload
Soduko solver implemented in Oracle PL/SQL. Implements the most common reduction strategies. And then, very important, it contains a recursive procedure, makeGuess, which whill be invoked in those cases the reduction strategies turn out to be incomplete. The procedure makesGuess ensures that the solver is complete: If a solution exists, it will be found.